+++ src/protocols/MQTT.h
@@ -33,13 +33,13 @@ class MQTTClient {
     MQTTClient(TransportLayer* tl, uint16_t port = 1883);
     
     // basic methods
-    uint8_t connect(const char* host, const char* clientId, const char* userName = "", const char* password = "", uint16_t keepAlive = 60, bool cleanSession = true, const char* willTopic = "", const char* willMessage = "");
-    uint8_t disconnect();
-    uint8_t publish(const char* topic, const char* message);
-    uint8_t subscribe(const char* topicFilter);
-    uint8_t unsubscribe(const char* topicFilter);
-    uint8_t ping();
-    uint8_t check(void (*func)(const char*, const char*));
+    int16_t connect(const char* host, const char* clientId, const char* userName = "", const char* password = "", uint16_t keepAlive = 60, bool cleanSession = true, const char* willTopic = "", const char* willMessage = "");
+    int16_t disconnect();
+    int16_t publish(const char* topic, const char* message);
+    int16_t subscribe(const char* topicFilter);
+    int16_t unsubscribe(const char* topicFilter);
+    int16_t ping();
+    int16_t check(void (*func)(const char*, const char*));
     
   private:
     TransportLayer* _tl;
